Isolate retry logic
